Add decode_document_utf8_lossy
MongoDB has a bug where it can return bson responses that contain
strings with invalid UTF-8 encoding:
https://jira.mongodb.org/browse/SERVER-24007

This bug has been open since 2016 and no fix is in sight, so we need
a workaround in the `mongo_driver` crate. This commit adds a separate
decode function that allows you to get lossy UTF-8 conversion. We will
use this function when parsing error responses from Mongo.